pyproject.toml
[project]
name = "landing"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "djangorestframework>=0.0.1",
    "drf-spectacular",
    "psycopg2-binary>=2.9.11",
    "python-dotenv",
    "gunicorn",
    "djangorestframework-simplejwt>=5.5.1",
    "django-cors-headers>=4.9.0",
]

[tool.isort]
profile = "black"

[dependency-groups]
dev = [
    "ruff",
    "black",
    "pre-commit>=4.5.1",
]
